requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
langchain-openai
youtube-transcript-api
openai
